[Shoot] Prefactor: extract defaulting into dedicated file
